+++ docs/CHANGELOG.md
@@ -166,3 +166,10 @@ Also set up a reusable animation system for future features:
 - The problem is that the red box is drawn in ShopMenu which cover the render of Button. Originally, it was done
 so to cover the item logs.
 - Now, ShopMenu has two background rendering files.
+
+### v0.14.2 - 2026-05-15
+- DayEndstate transition is changed from 'resume' to 'clear and push'
+- Now all customer will leave when the day ends or the shop closes.
+
+### v0.15.0 - 2026-05-17
+- when item is ready the bubby is now appear above the machine/plate